requirements.txt
numpy
pandas
faker
psycopg2-binary
sqlalchemy
fastapi
uvicorn[standard]
pydantic
requests
schedule
